refactor: change file extensions
