+++ src/manager/utils/doBye.mjs
@@ -1,4 +1,4 @@
-import { dictionary as dict } from '../../costants.js';
+import { dictionary as dict } from '../../costants.mjs';
 
 export const doBye = (username) => {
   console.log(dict.GET_BYE_TEXT(username));
