+++ src/manager/utils/getInvitationText.mjs
@@ -1,5 +1,5 @@
 import path from 'path';
-import { dictionary as dict } from '../../costants.js';
+import { dictionary as dict } from '../../costants.mjs';
 
 export const getInvitationText = (pathObject) =>
   `${dict.GET_CURRENT_DIR_TEXT(path.format(pathObject))}\n${
