+++ src/manager/utils/showErrorMsgOnExit.mjs
@@ -1,4 +1,4 @@
-import { dictionary as dict } from '../../costants.js';
+import { dictionary as dict } from '../../costants.mjs';
 
 export const showErrorMsgOnExit = () => {
   console.log(dict.ERROR_MSG_ON_EXIT);
